# Reduce allocations in Typer