I'm measuring a 2.7% reduction in allocation rate (`gc.alloc.rate.norm`) in `-Ystop-after:typer`. There might be a small speedup as a result of this,
but that reading is noisier than the allocation profiling.

Before:

```
 for sha in 1314b45 98b9a6c; do sbt 'set scalaVersion in compilation := "2.12.9-bin-'$sha'-SNAPSHOT"' 'hot -psource=scalap  -pextraArgs=-Ystop-after:typer -f3 -prof gc -jvmArgs -XX:MaxInlineLevel=18'; done
...

info] Benchmark                                                    (corpusPath)  (corpusVersion)         (extraArgs)  (resident)               (scalaVersion)  (source)    Mode   Cnt         Score         Error   Units
[info] HotScalacBenchmark.compile                                      ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample  1309       231.710 ±       0.571   ms/op
[info] HotScalacBenchmark.compile:compile·p0.00                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             223.085                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.50                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             229.638                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.90                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             240.386                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.95                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             244.318                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.99                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             254.568                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.999                       ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             271.167                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.9999                      ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             272.630                 ms/op
[info] HotScalacBenchmark.compile:compile·p1.00                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample             272.630                 ms/op
[info] HotScalacBenchmark.compile:·gc.alloc.rate                       ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30       378.301 ±       2.123  MB/sec
[info] HotScalacBenchmark.compile:·gc.alloc.rate.norm                  ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30  96551476.213 ±  195846.399    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space              ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30       379.304 ±      10.312  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm         ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30  96808470.750 ± 2608853.743    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space          ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30         2.729 ±       0.602  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm     ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30    695981.651 ±  152543.882    B/op
[info] HotScalacBenchmark.compile:·gc.count                            ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30       344.000                counts
[info] HotScalacBenchmark.compile:·gc.time                             ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-1314b45-SNAPSHOT    scalap  sample    30      2816.000                    ms
```

After:
```
[info] HotScalacBenchmark.compile                                      ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample  1324       228.940 ±       0.575   ms/op
[info] HotScalacBenchmark.compile:compile·p0.00                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             221.512                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.50                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             227.017                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.90                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             238.027                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.95                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             242.418                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.99                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             251.199                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.999                       ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             267.498                 ms/op
[info] HotScalacBenchmark.compile:compile·p0.9999                      ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             268.435                 ms/op
[info] HotScalacBenchmark.compile:compile·p1.00                        ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample             268.435                 ms/op
[info] HotScalacBenchmark.compile:·gc.alloc.rate                       ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30       376.478 ±       2.682  MB/sec
[info] HotScalacBenchmark.compile:·gc.alloc.rate.norm                  ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30  94957705.639 ±  127928.287    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space              ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30       374.880 ±      18.140  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm         ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30  94557225.117 ± 4515558.175    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space          ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30         1.849 ±       0.921  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm     ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30    468210.635 ±  232583.244    B/op
[info] HotScalacBenchmark.compile:·gc.count                            ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30       269.000                counts
[info] HotScalacBenchmark.compile:·gc.time                             ../corpus          a8c43dc  -Ystop-after:typer       false  2.12.9-bin-98b9a6c-SNAPSHOT    scalap  sample    30      2429.000                    ms
[success] Total time: 652 s, completed 15/08/2019 11:37:04 AM
```