@SethTisue 2019-08-30
“Has it come to this?” :D https://twitter.com/not_xuwei_k/status/1167445715267854336 😜